print.confint {MKmisc}	R Documentation

Print Method for Confidence Intervals

Description

Printing objects of class "confint" by a simple print method.

Usage

  ## S3 method for class 'confint'
print(x, digits = getOption("digits"), prefix = "\t", ...)

Arguments

x
object of class "confint".

digits
number of significant digits to be used.

prefix
string, passed to strwrap for displaying the method component of the mpe.test object.

...
further arguments to be passed to or from methods.

Details

A confint object is just a named list of confidence intervals and respective (point) estimates.

Value

the argument x, invisibly, as for all print methods.

See Also

print.power.htest

Examples

x <- rnorm(20)
(CI <- normCI(x))
print(CI, digits = 3)

[Package MKmisc version 1.9 Index]