rrCI {MKmisc}	R Documentation

Compute Approximate Confidence Interval for RR.

Description

The function computes an approximate confidence interval for the relative risk (RR).

Usage

rrCI(a, b, c, d, conf.level = 0.95)

Arguments

a
integer: events in exposed group.

b
integer: non-events in exposed group.

c
integer: events in non-exposed group.

d
integer: non-events in non-exposed group.

conf.level
numeric: confidence level

Details

The function computes an approximate confidence interval for the relative risk (RR) based on the normal approximation; see Jewell (2004).

Value

A list with class "confint" containing the following components:

estimate
the estimated relative risk.

conf.int
a confidence interval for the relative risk.

Author(s)

Matthias Kohl Matthias.Kohl@stamats.de

References

Jewell, Nicholas P. (2004). Statistics for epidemiology. Chapman & Hall/CRC.

Relative risk. (2016, November 4). In Wikipedia, The Free Encyclopedia. Retrieved 19:58, November 4, 2016, from https://en.wikipedia.org/w/index.php?title=Relative_risk&oldid=747857409

Examples

## See worked example in Wikipedia
rrCI(a = 15, b = 135, c = 100, d = 150)
rrCI(a = 75, b = 75, c = 100, d = 150)

[Package MKmisc version 1.9 Index]