simCorVars {MKmisc}	R Documentation

Simulate correlated variables.

Description

The function simulates a pair of correlated variables.

Usage

simCorVars(n, r, plot = TRUE)

Arguments

n
integer: sample size.

r
numeric: correlation.

plot
logical: generate scatter plot of the variables.

Details

The function is mainly for teaching purposes and simulates n observations from a pair of normal distributed variables with correlation r.

By specifying plot = TRUE a scatter plot of the data is generated.

Value

data.frame with entries Var1 and Var2

Author(s)

Matthias Kohl Matthias.Kohl@stamats.de

Examples

res <- simCorVars(n = 100, r = 0.8)
cor(res$Var1, res$Var2)

[Package MKmisc version 1.9 Index]