thyroid {MKmisc}	R Documentation

Plot TSH, fT3 and fT4 with respect to reference range.

Description

The function computes and plots TSH, fT3 and fT4 values with respect to the provided reference range.

Usage

thyroid(TSH, fT3, fT4, TSHref, fT3ref, fT4ref)

Arguments

TSH
numeric vector of length 1: measured TSH concentration.

fT3
numeric vector of length 1: measured fT3 concentration.

fT4
numeric vector of length 1: measured fT4 concentration.

TSHref
numeric vector of length 2: reference range TSH.

fT3ref
numeric vector of length 2: reference range fT3.

fT4ref
numeric vector of length 2: reference range fT4.

Details

A simple function that computes the relative values of the measured values with respect to the provided reference range and visualizes the values using a barplot. Relative values between 40% and 60% are marked as O.K..

Value

Invisible data.frame with the relative values.

Author(s)

Matthias Kohl Matthias.Kohl@stamats.de

Examples

thyroid(TSH = 1.5, fT3 = 2.5, fT4 = 14, TSHref = c(0.2, 3.0),
        fT3ref = c(1.7, 4.2), fT4ref = c(7.6, 15.0))

[Package MKmisc version 1.9 Index]