twoWayAnova {MKmisc}	R Documentation

A function for Analysis of Variance

Description

This function is a slight modification of function Anova of package "genefilter".

Usage

twoWayAnova(cov1, cov2, interaction, na.rm = TRUE)

Arguments

cov1
The first covariate. It must have length equal to the number of columns of the array that the result of twoWayAnova will be applied to.

cov2
The second covariate. It must have length equal to the number of columns of the array that the result of twoWayAnova will be applied to.

interaction
logical, should interaction be considered

na.rm
a logical value indicating whether 'NA' values should be stripped before the computation proceeds.

Details

The function returned by twoWayAnova uses lm to fit a linear model of the form lm(x ~ cov1*cov2), where x is the set of gene expressions. The F statistics for the main effects and the interaction are computed and the corresponding p-values are returned.

Value

twoWayAnova returns a function with bindings for cov1 and cov2that will perform a two-way ANOVA.

Note

A first version of this function appeared in package SLmisc.

Author(s)

Matthias Kohl Matthias.Kohl@stamats.de

References

R. Gentleman, V. Carey, W. Huber and F. Hahne (2006). genefilter: methods for filtering genes from microarray experiments. R package version 1.13.7.

See Also

Anova

Examples

set.seed(123)
af1 <- twoWayAnova(c(rep(1,6),rep(2,6)), rep(c(rep(1,3), rep(2,3)), 2))
af2 <- twoWayAnova(c(rep(1,6),rep(2,6)), rep(c(rep(1,3), rep(2,3)), 2), 
                   interaction = FALSE)
x <- matrix(rnorm(12*10), nrow = 10)
apply(x, 1, af1)
apply(x, 1, af2)

[Package MKmisc version 1.9 Index]